python: Update wrapping function to support kit
The previous commit allows python module to be successfully
imported when running vtkpython from a build tree. This commit
improved even further by updating the wrapping function to avoid
manually creating the Init.cxx file for a kit.

Instead, the function "vtk_wrap_python"
and "vtk_add_python_wrapping" have been updated to understand a list
of module as input.

Similarly to the approach used in the previous commit, it has been
verified that the content of the vtk module was the expected
one for a build and install tree with or without the
VTK_ENABLE_KITS option enabled.

Cherry-picked from:
https://github.com/Slicer/VTK/commit/3e3ca05