COMP: VTK and ITK H5 build program collision
When both VTK and ITK are built with the same
CMAKE_RUNTIME_OUTPUT_DIRECTORY settings, the
H5detect and H5make_libsettings programs are
generated by both VTK and ITK.  These are
intended to be internal build configuration
programs that determine the capabilities
of the local compiler and OS.

When VTK and ITK diverge in their version of
HDF5, conficts occur that cause difficult to
debug compiler errors.

By adding a prefix here
the program name conflict can be avoided.