mpi4py: Fix static/shared module mixup
The PYTHON_ADD_MODULE macro handles whether or not to build the python
module as a static library or an so.  By explicitly calling the macro
with the MODULE argument, the logic inside gets borked and always build
it as shared.  This removes the MODULE argument from the macro call and
let's the macros logic determing apropriately whether to build as static
or shared.