Fix PExodusIIWriter called without GlobalMPController.
If you enable the Module_vtkIOParallelExodus option, the object factory
for a vtkExodusIIWriter will return a vtkPExodusIIWriter.
However this parallel version suppose a parallel use of the writer
and a global MultiProcessController to be set which is not the case
if you use the writer in serial.
This patch ensure that the writer will not crash and will have the same
behaviour than the (hidden) serial writer.