python: fix vtkOpenGLKitPythonD linker error
This commit fixes the following linker error when VTK_ENABLE_KITS is ON and
VTK_RENDERING_BACKEND is OpenGL2:

    Linking CXX shared library ../../lib/libvtkOpenGLKitPython27D-7.1.so
    CMakeFiles/vtkOpenGLKitPythonD.dir/vtkOpenGLMoleculeMapperPython.cxx.o: In function `PyvtkOpenGLMoleculeMapper_ClassNew':
    /home/msmolens/dev/VTK-build/Wrapping/Python/vtkOpenGLMoleculeMapperPython.cxx:307: undefined reference to `PyvtkMoleculeMapper_ClassNew'

The order in which kits and modules are wrapped now satisfies dependencies
between the kits and modules. This change ensures that all the targets for
dependent modules exist when vtk_add_python_wrapping_library is called.

Fixes #15255