Modernized vtkSortDataArray, threaded the sort
Cleaned up and modernized vtkSortDataArray. Using std::sort and
threaded the sort operation (with vtkSMPTools). Removed static
variable that made it non-thread-safe. Added the ability to sort
in ascending as well as descending order. Expanded and improved
the test TestSortDataArray.

Note that the code handles vtkStdString and vtkVariant.

Note also: various types of arrays, including the vtkIdList, needed
to be expanded so the SetArray() (or equivalent operation to
specify the under-the-hood array to use). This is so the sort
can swap out the data once the data is shuffled.