python: Add adapter modules when VTK_ENABLE_KITS is ON
Some Python scripts explicitly import from named submodules as opposed to from
the top-level vtk module. When VTK_ENABLE_KITS is ON those modules might not
exist. This results in ImportErrors, like the one shown below when running the
vtkFiltersCorePython-pointsPrecisions test:

    Traceback (most recent call last):
      File "C:/dev/VTK/Filters/Core/Testing/Python/pointsPrecisions.py", line 3, in <module>
	from vtk.vtkCommonCore import vtkPoints, vtkDoubleArray, vtkIdList
    ImportError: No module named vtkCommonCore

This commit adds adapter modules to allow such scripts to run unmodified when
VTK_ENABLE_KITS is ON.