Fixed out-of-bounds vector<int> access
A vector<int> was being resized too small.
Changed resize() to reserve(), which I believe is what was
intended.  Thanks to Andrew Bauer to getting to the
heart of the solution.

Also, instead of passing int*, pass source vector<int>.
This allows asserting its size to detect buffer overruns.

Fixes the vtkFiltersCoreCxx-TestStructuredGridAppend
test that provokes an overrun, which I found with ASan
(address sanitizer).