Reduce depends of OpenGL2
This patch makes it so that OpenGL2 depends on fewer other modules
Previously it required > 40 modules to build. Now it can be done
for 19 with only minor changes. This patch also required adding
a capability so that implementations can be optional otherwise
gl2ps would always bring in a bunch of other modules when opengl2
was built.