python: add backwards compatibility for wrapping functions
The commit titled "python: split wrapping functions" breaks backwards
compatibility for the vtk_add_python_wrapping() function. Client code that has
not been updated would result in errors like:

    Function invoked with incorrect arguments for function named:
    vtk_add_python_wrapping

A warning message is printed when this function is called in the legacy manner.
Note that the CMake message function's DEPRECATED option is not used so that
CMake < 3.0 can be used.

To upgrade client code, replace calls like:

    vtk_add_python_wrapping(${module})

with:

    vtk_add_python_wrapping(${module} module_srcs)
    vtk_add_python_wrapping_library(${module} module_srcs ${module})