COMP: The Apple GNU compiler/linker has issues with libprog
After the recent upgrade to libproj, the GNU linker reports undefined
symbols while building vtklibproj and vtkGeovisCore.

This is apprently a bug in this old GNU compiler/linker. This patch
works around the problem by building booth libvtkproj and
vtkGeovisCore as static libraries.

This patch has been tested with this compiler:
  i686-apple-darwin11-llvm-g++-4.2 (GCC) 4.2.1
  (Based on Apple Inc. build 5658) (LLVM build 2336.1.00)