Measure depth complexity to determine required passes.
This patch adds a stencil buffer to the initial geometry pass of the
dual depth peeling algorithm which is used to count how many fragments
per pixel will need to be handled.

This buffer is asynchronously transferred to system memory while
the peeling algorithm continues running for up to 5 passes. Once
available, the buffer is inspected to identify the depth complexity of
the scene and limit the number of passes.

This greatly reduces the need to wait for occlusion query results, since
we know a-priori how many passes are needed to reach the occlusion
critera.

(update) After some testing, it looks like the stencil buffer is double
counting some fragments, leading to an overestimation in the number of
passes required. I'll keep the last number of passes and check occlusion
at both places.