adding a faster, selectable, text mapper
It does not comply with ALL of the controls that you can get
from the standard text mapper and TextProperty, but it does
support most of them including font choice, frames and highlight
color.

The basic idea is to make a single texture that has each
character that is to be drawn, and then use shaders to build
words at the right place and size character by character.

Cheers to Allison Vacanti for making the character alignment
pixel-perfect.