vtkOpenFOAMReader: Support more "nonuniform" Lists
In OpenFOAM boundary condition files, lists associated with the geometry
(cells or faces) have `uniform` or `nonuniform` attribute to indicate
whether a single value is assigned to all faces or there is one value
for each face. However, boundary conditions can contain all other kinds
of lists (not geometrically motivated) as well, for example widths of
thermal resistance layers, positions of some averaging rings and many
others, particularly in the case of custom boundary conditions.

In this commit the reading of the boundary condition dictionary is
adjusted so that after attempting to read `uniform` and `nonuniform`
lists, the reader now also attempts to read a list not introduced with
such keyword. This generalizes a recent modification related to
`List<scalar>`.