Extract Selected Cells by ID Performance
Added a new filter to select cells by id number more efficiently when the
selection list can be defined by a dense boolean array of in/out flags
spanning an id range of [0->n].

Extended the PreserveTopology feature to allow multiple selection filters
to progressively filter in sequence.

Added the PreserveTopology feature to the surface filter to reduce the
number of dataset copies required in pipelines which filter cells before
generating a surface.

PreserveTopology will continue to set an array into the output data, but
will no longer also call SetScalars as this interferes with the color
lookup tables and assigning scalar property values to the dataset.